requirements.txt
cleverbotfree==2.3.3
imgurpython==1.1.7
PyGithub==1.55
requests==2.27.1
simple_image_download==0.2
Unidecode==1.3.2
websocket_client==1.3.1
